resources/python/sources/sql_cdc/requirements.txt
quixstreams==2.9.0
pyodbc
requests
SQLAlchemy
python-dotenv